@@ -7,10 +7,9 @@ a release version. Release versions are git tagged with the version.
 """
 
 name = "lmu"
-version_info = (0, 1, 1)  # (major, minor, patch)
+version_info = (0, 2, 0)  # (major, minor, patch)
 dev = 0
 
-version = "{v}{dev}".format(
-    v=".".join(str(v) for v in version_info),
-    dev=(".dev%d" % dev) if dev is not None else "",
+version = (
+    f"{'.'.join(str(v) for v in version_info)}{f'.dev{dev}' if dev is not None else ''}"
 )
